requirements.txt
Django==4.2.5
line-bot-sdk==3.2.0
python-dotenv==1.0.0